requirements.txt
matplotlib==3.3.3
streamlit==1.4
numpy==1.19.5
pandas==1.1.5
scikit_learn==0.24.1
scipy==1.6.3
seaborn==0.11.1
xlrd==1.2.0
openpyxl==3.0.7
plotly_express
datetime
folium==0.12.1
streamlit_folium==0.4.0
xgboost==1.4.2
google-cloud-bigquery==2.26.0
ortools==9.3.10497